Consolidate dread command
