--- inodaqv2/command_dread.h
@@ -1,6 +1,0 @@
-#pragma once
-
-namespace Command
-{
-void command_dread();
-}
