+++ inodaqv2/commands.hpp
@@ -4,3 +4,4 @@
 
 void command_aread();
 void command_dig(const ::String &command);
+void command_dread();
